#23 move lib and utils to commons folder
