+++ src/app/components/Popup/Popup.js
@@ -2,7 +2,7 @@ import React from "react";
 import { createHashHistory } from "history";
 import { HashRouter, Switch, Route } from "react-router-dom";
 
-import utils from "../../../lib/utils";
+import utils from "../../../common/lib/utils";
 import Home from "../../pages/home";
 import Unlock from "../../pages/unlock";
 import Loading from "../../pages/loading";
