+++ src/app/components/Welcome/Welcome.js
@@ -1,9 +1,9 @@
 import React from "react";
 
-import utils from "../../../lib/utils";
-import Accounts from "../../../lib/accounts";
-import Allowances from "../../../lib/allowances";
-import Settings from "../../../lib/settings";
+import utils from "../../../common/lib/utils";
+import Accounts from "../../../common/lib/accounts";
+import Allowances from "../../../common/lib/allowances";
+import Settings from "../../../common/lib/settings";
 
 class Welcome extends React.Component {
   constructor(props) {
