+++ src/common/lib/connectors/base.js
