+++ src/common/lib/connectors/lnd.js
