+++ src/common/lib/connectors/native.js
