+++ src/common/lib/crypto.js
