+++ src/common/lib/msg.js
