+++ src/common/lib/settings.js
